pyproject.toml
[project]
name = "latex-resume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-api-python-client>=2.176.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
]
